+++ Interfaces/Repository/IDepartmentRepository.cs
@@ -5,7 +5,7 @@ namespace Interfaces.Repository
     /// <summary>
     /// Department Repository Interface
     /// </summary>
-    public interface IDepartmentRepository : IBaseRepository<Department, int>
+    public interface IDepartmentRepository : IBaseRepository<Department, long>
     {
 
     }
