+++ Interfaces/Repository/IRegionRepository.cs
@@ -5,7 +5,7 @@ namespace Interfaces.Repository
     /// <summary>
     /// Region Repository Interface
     /// </summary>
-    public interface IRegionRepository : IBaseRepository<Region, int>
+    public interface IRegionRepository : IBaseRepository<Region, long>
     {
     }
 }
